Add files via upload
This is a  basic ACMshell , a command-line interpreter or a command language interpreter that provides a user interface for interacting with the operating system. It serves as a primary interface for executing commands, managing processes, and accessing various system resources. The shell acts as a bridge between the user and the underlying Unix-like operating system, facilitating the execution of commands and the management of system resources.

 ACMShell provide facilities to redirect input and output streams. Users can redirect standard input (stdin), standard output (stdout), and standard error (stderr) to files or to other commands, enabling flexible data processing and file manipulation.

It is composed of some Built-in Commands which include a set of built-in commands that are executed directly by the shell itself, without invoking external programs. Built-in commands provide essential functionality and often offer performance advantages compared to external commands.